Remove processError in favor of just triggering event
Previously the error message was hard coded and it
was sent to all the same selectors that the developer
likely expected to change on success.

Furthermore, this is handling edge case scenarios,
namely 404 and 500 errors and this seems best left
up to the site developer how it should be handled.