Start of rewrite
We start with a new test framework with Jasmine and Karma, followed
by the start of exposing things outside of the anonymous blocks so
we can test better.

Lots of work to port the rest of the tests. This commit only ports
a handful so that we can establish a beach head on the new approach.

We'll focus on other items in this branch as part of #89